refactor: tighten github transport boundary
