+++ adl/src/cli/pr_cmd/github/tests.rs
@@ -4,7 +4,7 @@ use std::fs;
 use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
 use std::thread;
-use std::time::Duration;
+use std::time::{Duration, Instant};
 use tiny_http::{Header, Response, Server};
 
 fn env_lock() -> std::sync::MutexGuard<'static, ()> {
